+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Node path to hover
+- Property DTS native type and binding type to hover
 - LPS API to get actions for a location:
   - Generate C Identifers macros for nodes and properties
   - Generate node path